refactor: separate Skills into standalone settings section per feedback
- Moved Skills from tabbed interface to standalone settings section
- Restored SlashCommandsSettings to original non-tabbed state
- Created new SkillsSettings component wrapping SkillsTab
- Updated SettingsView to include both sections separately
- Updated all i18n translations (18 locales) for new section structure
- Removed CommandsAndSkillsSettings component and its tests
- Added SkillsSettings test suite
- All tests passing (42/42)

Addresses feedback from Bruno Bergher requesting Skills be separated
from slash commands into their own section due to complexity differences.
