+++ webview-ui/src/components/settings/SettingsView.tsx
@@ -28,6 +28,7 @@ import {
 	Server,
 	Users2,
 	ArrowLeft,
+	Sparkles,
 } from "lucide-react"
 
 import {
@@ -75,7 +76,8 @@ import { LanguageSettings } from "./LanguageSettings"
 import { About } from "./About"
 import { Section } from "./Section"
 import PromptsSettings from "./PromptsSettings"
-import { CommandsAndSkillsSettings } from "./CommandsAndSkillsSettings"
+import { SlashCommandsSettings } from "./SlashCommandsSettings"
+import { SkillsSettings } from "./SkillsSettings"
 import { UISettings } from "./UISettings"
 import ModesView from "../modes/ModesView"
 import McpView from "../mcp/McpView"
@@ -97,6 +99,7 @@ export const sectionNames = [
 	"providers",
 	"autoApprove",
 	"slashCommands",
+	"skills",
 	"browser",
 	"checkpoints",
 	"notifications",
@@ -525,6 +528,7 @@ const SettingsView = forwardRef<SettingsViewRef, SettingsViewProps>(({ onDone, t
 			{ id: "mcp", icon: Server },
 			{ id: "autoApprove", icon: CheckCheck },
 			{ id: "slashCommands", icon: SquareSlash },
+			{ id: "skills", icon: Sparkles },
 			{ id: "browser", icon: SquareMousePointer },
 			{ id: "checkpoints", icon: GitBranch },
 			{ id: "notifications", icon: Bell },
@@ -811,8 +815,11 @@ const SettingsView = forwardRef<SettingsViewRef, SettingsViewProps>(({ onDone, t
 							/>
 						)}
 
-					{/* Commands & Skills Section */}
-					{renderTab === "slashCommands" && <CommandsAndSkillsSettings />}
+					{/* Slash Commands Section */}
+					{renderTab === "slashCommands" && <SlashCommandsSettings />}
+
+					{/* Skills Section */}
+					{renderTab === "skills" && <SkillsSettings />}
 
 						{/* Browser Section */}
 						{renderTab === "browser" && (
