+++ webview-ui/src/i18n/locales/en/settings.json
@@ -35,8 +35,8 @@
 		"notifications": "Notifications",
 		"contextManagement": "Context",
 		"terminal": "Terminal",
-		"slashCommands": "Commands & Skills",
-		"commandsAndSkills": "Commands & Skills",
+		"slashCommands": "Slash Commands",
+		"skills": "Skills",
 		"prompts": "Prompts",
 		"ui": "UI",
 		"experimental": "Experimental",
