+++ webview-ui/src/i18n/locales/nl/settings.json
@@ -40,8 +40,7 @@
 		"ui": "UI",
 		"experimental": "Experimenteel",
 		"language": "Taal",
-		"about": "Over Roo Code",
-		"commandsAndSkills": "Commando's & Vaardigheden"
+		"about": "Over Roo Code"
 	},
 	"about": {
 		"bugReport": {
@@ -1012,10 +1011,6 @@
 			"label": "Vereist {{primaryMod}}+Enter om berichten te versturen",
 			"description": "Wanneer ingeschakeld, moet je {{primaryMod}}+Enter indrukken om berichten te versturen in plaats van alleen Enter"
 		}
-	},
-	"commandsAndSkills": {
-		"tabCommands": "Slash-commando's",
-		"tabSkills": "Vaardigheden"
 	},
 	"skills": {
 		"global": "Globale vaardigheden",
