+++ webview-ui/src/i18n/locales/pt-BR/settings.json
@@ -40,8 +40,7 @@
 		"ui": "UI",
 		"experimental": "Experimental",
 		"language": "Idioma",
-		"about": "Sobre",
-		"commandsAndSkills": "Comandos e Habilidades"
+		"about": "Sobre"
 	},
 	"about": {
 		"bugReport": {
@@ -1012,10 +1011,6 @@
 			"label": "Requer {{primaryMod}}+Enter para enviar mensagens",
 			"description": "Quando ativado, você deve pressionar {{primaryMod}}+Enter para enviar mensagens em vez de apenas Enter"
 		}
-	},
-	"commandsAndSkills": {
-		"tabCommands": "Comandos de barra",
-		"tabSkills": "Habilidades"
 	},
 	"skills": {
 		"global": "Habilidades globais",
